routine(phase-6.3): parse SRAT x2APIC processor affinity (type 2)
srat::cpu_affinities now covers Processor Local x2APIC Affinity (type 2,
32-bit x2APIC ID at +8, proximity domain at +4) in addition to type 0,
and CpuAffinity.apic_id is widened to u32 so CPUs with APIC IDs above 255
get their NUMA node. build_device_tree's MADT-APIC-ID -> SRAT-node match
uses the u32 id directly. Closes the x2APIC gap noted when the device
tree landed. Unit-tested (enabled x2APIC id 300 -> domain 5; disabled
skipped).

Built + tested on Linux (WSL, 10 srat tests) and Windows-native; full
clippy --all-targets clean.

+++ enlil-devices/src/acpi/discover.rs
@@ -312,7 +312,7 @@ pub fn build_device_tree(mem: &[u8], rsdp_gpa: u64) -> HostDeviceTree {
             apic_id,
             proximity_domain: affinities
                 .iter()
-                .find(|a| u32::from(a.apic_id) == apic_id)
+                .find(|a| a.apic_id == apic_id)
                 .map(|a| a.proximity_domain),
         })
         .collect();
